organize lib into dirs
